Implement node module resolution algo in VimScript
Most of node's module resolution algorithm is now implemented in Further
with the exception of a few small items:
- No module builtin support (same as before)
- No support for $NODE_PATH (fix coming soon)
- `pkg.main` fallback isn't fully implemented (next on agenda)

This has several advantages, but the main advantage is first-class support
for more extensions, like `.tsx` and `.mjs`. If the extension of choice
isn't listed you can extend it with `further#extensions`.