Consider package.json for relative directories
Whenever you import `../<directory>`, further should handle the
possibility that it contains a package.json file. This sort of thing
probably only happens if you're in a monorepo and importing another
package without declaring it in the dependencies (you bad person you).

The only exception to this rule is when resolving `pkg.main`. If the
result is a directory with another package.json, ignore it and look for
`index`.