Mention Plug'n'Play
If you've found this commit, you might be wondering "but why dude" and
I'm here to say "well golly it's just darn hard". Please read that in
a funny accent.

The goal of further.vim is to resolve JavaScript modules, and for it to
execute reasonably, it must be fast. Node's startup time is murder, and
if I have to invoke a JS script to resolve a module, that rounds out to
about a second of blocking overhead. So I implemented it in Vimscript,
increasing speed by one order of magnitude.

I'm not willing to pay the performance price of invoking `.pnp.js`. I'd
rather implement it in a new plugin versus native in further.vim.

Alternatively, I could expose a neovim remote plugin integration that
maintains a persistent runtime and exposes PnP resolution functions, but
it wouldn't work with vim 8, it would have to be async (blocking is
actually desirable), it would require effort to reload the environment
between lockfile changes, and even after resolution, I still need to
extract library zip files into a temp directory for inspection.

Is all that worth it? And does it belong in further.vim? I don't think
so, and that's why I've said "No".